+++ src/app/model/events/event.ts
@@ -2,7 +2,7 @@ import { AggregateTypeEnum } from "../enums/aggregateType.enum";
 import { EventTypeEnum } from "../enums/eventType.enum";
 import { EventDataModel } from "./eventDataModel";
 import { EventStatusEnum } from "../enums/eventStatus.enum";
-import { EventVisitor } from "../../services/notifications/visitors/eventVisitor";
+import { EventVisitor } from "../../services/eventvisitors/eventVisitor";
 
 export interface Event {
   eventId: string;
