+++ src/app/model/events/privateEvent.model.spec.ts
@@ -1,4 +1,4 @@
-import { EventVisitor } from "../../services/notifications/visitors/eventVisitor";
+import { EventVisitor } from "../../services/eventvisitors/eventVisitor";
 import { AggregateTypeEnum } from "../enums/aggregateType.enum";
 import { EventTypeEnum } from "../enums/eventType.enum";
 import { EventStatusEnum } from "../enums/eventStatus.enum";
