+++ src/app/services/eventhandlers/eventDataValidators.ts
@@ -1,7 +1,7 @@
-import { EventDataModel } from "../../../model/events/eventDataModel";
-import { MemberModel } from "../../../model/member.model";
-import { ErrorDataModel } from "../../../model/errorData.model";
-import { GroupModel } from "../../../model/group.model";
+import { EventDataModel } from "../../model/events/eventDataModel";
+import { MemberModel } from "../../model/member.model";
+import { ErrorDataModel } from "../../model/errorData.model";
+import { GroupModel } from "../../model/group.model";
 
 export function isEventDataGroupModel(
   eventData: EventDataModel,
