+++ src/app/services/eventhandlers/eventHandler.ts
@@ -1,0 +1,7 @@
+import { PublicEventModel } from "../../model/events/publicEvent.model";
+import { PrivateEventModel } from "../../model/events/privateEvent.model";
+
+export interface EventHandler {
+  handlePublicEvent(event: PublicEventModel): void;
+  handlePrivateEvent(event: PrivateEventModel): void;
+}
