+++ src/app/services/eventhandlers/groupEventHandlers/joinGroup.handler.spec.ts
@@ -1,13 +1,13 @@
 import { JoinGroupHandler } from "./joinGroup.handler";
-import { UserService } from "../../../user/user.service";
-import { NotificationService } from "../../notification.service";
-import { PublicEventModel } from "../../../../model/events/publicEvent.model";
-import { PrivateEventModel } from "../../../../model/events/privateEvent.model";
+import { UserService } from "../../user/user.service";
+import { NotificationService } from "../../notifications/notification.service";
+import { PublicEventModel } from "../../../model/events/publicEvent.model";
+import { PrivateEventModel } from "../../../model/events/privateEvent.model";
 import { TestBed } from "@angular/core/testing";
-import { EventStatusEnum } from "../../../../model/enums/eventStatus.enum";
-import { MemberModel } from "../../../../model/member.model";
-import { MemberStatusEnum } from "../../../../model/enums/memberStatus.enum";
-import { ErrorDataModel } from "../../../../model/errorData.model";
+import { EventStatusEnum } from "../../../model/enums/eventStatus.enum";
+import { MemberModel } from "../../../model/member.model";
+import { MemberStatusEnum } from "../../../model/enums/memberStatus.enum";
+import { ErrorDataModel } from "../../../model/errorData.model";
 
 describe("JoinGroupHandler", () => {
   let joinGroupHandler: JoinGroupHandler;
