+++ src/app/services/eventvisitors/eventVisitor.ts
@@ -1,0 +1,7 @@
+import { PublicEventModel } from "../../model/events/publicEvent.model";
+import { PrivateEventModel } from "../../model/events/privateEvent.model";
+
+export interface EventVisitor {
+  visitPrivateEvent(event: PrivateEventModel): void;
+  visitPublicEvent(event: PublicEventModel): void;
+}
