+++ src/app/services/eventvisitors/group/groupEvent.visitor.ts
@@ -1,11 +1,11 @@
 import { EventVisitor } from "../eventVisitor";
-import { PublicEventModel } from "../../../../model/events/publicEvent.model";
+import { PublicEventModel } from "../../../model/events/publicEvent.model";
 import { Injectable } from "@angular/core";
 import { GroupUpdatedHandler } from "../../eventhandlers/groupEventHandlers/groupUpdated.handler";
 import { JoinGroupHandler } from "../../eventhandlers/groupEventHandlers/joinGroup.handler";
 import { LeaveGroupHandler } from "../../eventhandlers/groupEventHandlers/leaveGroup.handler";
-import { EventTypeEnum } from "../../../../model/enums/eventType.enum";
-import { PrivateEventModel } from "../../../../model/events/privateEvent.model";
+import { EventTypeEnum } from "../../../model/enums/eventType.enum";
+import { PrivateEventModel } from "../../../model/events/privateEvent.model";
 
 @Injectable({
   providedIn: "root",
